refactor: move polylogue/verification manifest models into devtools
Problem: polylogue/verification/{__init__.py, manifests/__init__.py,
manifests/models.py} had a single production importer,
devtools/verify_manifests.py:23. The package lived under the polylogue/
substrate but was purely a devtools-consumed set of Pydantic models for
docs/plans/*.yaml manifests, not archive/session domain logic
(polylogue-zu0kt).

What changed: moved manifests/models.py to devtools/manifest_models.py
(flat file, matching the existing devtools/benchmark_models.py-style
convention rather than a nested package, given its ~600-line size) and
updated the sole import site. Deleted the now-empty
polylogue/verification/ package. Updated TESTING.md's path reference to
the module. Regenerated docs/plans/topology-target.yaml since the module
move changes the topology surface. Confirmed via grep that no other file
imports polylogue.verification.

Verification: devtools test tests/unit/devtools/test_verify_manifests.py
(26 passed); devtools verify --quick (exit 0, includes mypy --strict,
layering check, and render all --check).
