--- polylogue/verification/__init__.py
@@ -1,1 +1,0 @@
-"""Verification infrastructure — typed manifests, anti-vacuity, coverage contracts."""
