--- polylogue/verification/manifests/__init__.py
@@ -1,1 +1,0 @@
-"""Pydantic models and validators for docs/plans/*.yaml manifests."""
